moved convertAmount to utilities.go
