+++ internal/common/shared/iigocommunications.go
@@ -106,8 +106,8 @@ func (c CommunicationFieldName) String() string {
 		"PresidentID",
 		"RuleName",
 		"RuleVoteResult",
-		"TaxDecision",
-		"AllocationDecision",
+		"Tax",
+		"Allocation",
 		"PardonClientID",
 		"PardonTier",
 		"SanctionAmount",
