+++ internal/server/iigointernal/executive.go
@@ -10,7 +10,6 @@ import (
 	"github.com/SOMAS2020/SOMAS2020/internal/common/shared"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/voting"
 	"github.com/pkg/errors"
-	"gonum.org/v1/gonum/mat"
 )
 
 type executive struct {
@@ -242,46 +241,6 @@ func (e *executive) incurServiceCharge(cost shared.Resources) bool {
 		e.gameState.IIGORolesBudget[shared.President] -= cost
 	}
 	return ok
-}
-
-// convertAmount takes the amount of tax/allocation and converts it into appropriate variable and rule ready to be sent to the client
-func convertAmount(amount shared.Resources, amountType conversionType) (rules.VariableValuePair, rules.RuleMatrix) {
-	var reqVar rules.VariableFieldName
-	var sentVar rules.VariableFieldName
-	var ruleVariables []rules.VariableFieldName
-	if amountType == tax {
-		reqVar, sentVar = rules.IslandTaxContribution, rules.ExpectedTaxContribution
-		// Rule in form IslandTaxContribution - ExpectedTaxContribution >= 0
-		ruleVariables = []rules.VariableFieldName{reqVar, sentVar}
-	} else if amountType == allocation {
-		reqVar, sentVar = rules.IslandAllocation, rules.ExpectedAllocation
-		// Rule in form ExpectedAllocation - IslandAllocation >= 0
-		ruleVariables = []rules.VariableFieldName{sentVar, reqVar}
-	}
-
-	v := []float64{1, -1, 0}
-	aux := []float64{2}
-
-	rowLength := len(ruleVariables) + 1
-	nrows := len(v) / rowLength
-
-	CoreMatrix := mat.NewDense(nrows, rowLength, v)
-	AuxiliaryVector := mat.NewVecDense(nrows, aux)
-
-	retRule := rules.RuleMatrix{
-		RuleName:          fmt.Sprintf("%s = %.2f", sentVar, amount),
-		RequiredVariables: ruleVariables,
-		ApplicableMatrix:  *CoreMatrix,
-		AuxiliaryVector:   *AuxiliaryVector,
-		Mutable:           false,
-	}
-
-	retVar := rules.VariableValuePair{
-		VariableName: sentVar,
-		Values:       []float64{float64(amount)},
-	}
-
-	return retVar, retRule
 }
 
 func (e *executive) sendTax(islandID shared.ClientID, taxAmount shared.Resources) {
